move trigger config back to producer, remove new method, add copy method
